moved generic member access into MemberContainerObject base class
