+++ nineml/base.py
@@ -1,4 +1,7 @@
 from itertools import chain
+from collections import defaultdict
+from nineml.exceptions import (
+    NineMLRuntimeError, NineMLInvalidElementTypeException)
 
 
 class BaseNineMLObject(object):
@@ -134,6 +137,144 @@ class TopLevelObject(object):
         Writes the top-level NineML object to file in XML.
         """
         nineml.write(self, fname)  # Calls nineml.document.Document.write
+
+
+class MemberContainerObject(object):
+    """
+    An abstract base class for handling the manipulation of member objects
+    (which are stored in dictionaries that can be detected by member type).
+
+    Deriving classes are expected to have the '_class_to_member' class
+    attribute
+    """
+
+    def __init__(self):
+        self._indices = defaultdict(dict)
+
+    def elements(self):
+        return chain(*(d.itervalues() for d in self._all_member_dicts))
+
+    def add(self, element):
+        dct = self._member_dict(element)
+        if element.name in dct:
+            raise NineMLRuntimeError(
+                "Could not add '{}' {} to component class as it clashes with "
+                "an existing element of the same name"
+                .format(element.name, type(element).__name__))
+        dct[element.name] = element
+
+    def remove(self, element, ignore_missing=False):
+        dct = self._member_dict(element)
+        try:
+            del dct[element.name]
+        except KeyError:
+            if not ignore_missing:
+                raise NineMLRuntimeError(
+                    "Could not remove '{}' from component class as it was not "
+                    "found in member dictionary".format(element.name))
+
+    def __getitem__(self, name):
+        for dct in self._all_member_dicts:
+            try:
+                return dct[name]
+            except KeyError:
+                pass
+        raise KeyError("'{}' was not found in '{}' component class"
+                       .format(name, self.name))
+
+    def __contains__(self, element):
+        """
+        Comprehensively checks whether the element belongs to this component
+        class or not. Useful for asserts and unit tests.
+        """
+        if isinstance(element, basestring):
+            try:
+                self[element]
+            except KeyError:
+                return False
+        else:
+            try:
+                dct = self._member_dict(element)
+                try:
+                    found = dct[element.name]
+                    assert(found == element)
+                    return True
+                except KeyError:
+                    return False
+            except NineMLInvalidElementTypeException:
+                return self._find_element(element)
+
+    @property
+    def _all_member_dicts(self):
+        return chain(*([m.itervalues() for m in self._class_to_member] +
+                       [m.itervalues()
+                        for m in self._main_block._class_to_member]))
+
+    @property
+    def elements(self):
+        return chain(*(d.itervalues() for d in self._all_member_dicts))
+
+    def index_of(self, element, key=None):
+        """
+        Returns the index of an element amongst others of its type. The indices
+        are generated on demand but then remembered to allow them to be
+        referred to again.
+
+        This function is meant to be useful in code-generation, where an
+        name of an element can be replaced with a unique integer value (and
+        referenced elsewhere in the code).
+        """
+        if key is None:
+            key = self._member_dict_name(element)
+        dct = self._indices[key]
+        try:
+            index = dct[element]
+        except KeyError:
+            # Get the first index ascending from 0 not in the set
+            try:
+                index = next(iter(sorted(
+                    set(xrange(len(dct)).difference(dct.itervalues())))))
+            except:
+                index = len(dct)
+            dct[element] = index
+        return index
+
+    def _member_dict_name(self, element):
+        # Try quick lookup by class type
+        try:
+            return self._class_to_member[type(element)]
+        except KeyError:
+            # For component classes with "main blocks" (to be removed in
+            # version 2.0)
+            try:
+                return self._main_block._class_to_member[type(element)]
+            except (AttributeError, KeyError):
+                pass
+        try:
+            return next(
+                d for cls, d in self._class_to_member.iteritems()
+                if isinstance(element, cls))
+        except StopIteration:
+            # For component classes with "main blocks" (to be removed in
+            # version 2.0)
+            try:
+                return next(
+                    d for cls, d in
+                    self._main_block._class_to_member.iteritems()
+                    if isinstance(element, cls))
+            except (AttributeError, StopIteration):
+                raise NineMLInvalidElementTypeException(
+                    "Could not get member dict for element of type "
+                    "'{}' to '{}' class"
+                    .format(element.__class__.__name__,
+                            self.__class__.__name__))
+
+    def _member_dict(self, element):
+        name = self._member_dict_name(element)
+        try:
+            return getattr(self, name)
+        except AttributeError:
+            return getattr(self._main_block, name)
 
 
 import nineml
